+++ RemEx.Android/app/src/main/java/com/clindsay94/remex/data/SettingsManager.kt
@@ -104,7 +104,7 @@ class SettingsManager(val context: Context) {
 
         data class RemoteDesktopPreferences(
                 val quality: Int = 50,
-                val targetFps: Int = 30,
+                val targetFps: Int = 120,
                 val scale: Float = 0.6f,
                 val directTouch: Boolean = false,
                 val pointerSpeed: Float = 1.0f,
@@ -274,7 +274,7 @@ class SettingsManager(val context: Context) {
                 context.dataStore.data.map { preferences ->
                         RemoteDesktopPreferences(
                                 quality = preferences[DESKTOP_QUALITY_KEY] ?: 50,
-                                targetFps = preferences[DESKTOP_TARGET_FPS_KEY] ?: 30,
+                                targetFps = preferences[DESKTOP_TARGET_FPS_KEY] ?: 120,
                                 scale = preferences[DESKTOP_SCALE_KEY] ?: 0.6f,
                                 directTouch = preferences[DESKTOP_DIRECT_TOUCH_KEY] ?: false,
                                 pointerSpeed = preferences[DESKTOP_POINTER_SPEED_KEY] ?: 1.0f,